InviteAndJoin.pl plugin for xchat
InviteAndJoin.pl plugin for xchat which waits for you to identify with NickServ
and then requests invites for all channels in a list.  When the invite is
recieved then the plugin will automatically join the channel for you.

To use put this in your "connect command":
load /path/to/InviteAndJoin.pl