New Usertag: [both]
This usertag will return the entire contents concatenated together if all of
the individual components contain something that is not whitespace.  Example:

Old method:
foo[if scratch bar] - [scratch bar][/if]

With this usertag:
foo[both] - [and][scratch bar][/both]

Note the difference that if [scratch bar] contains 0 then [both] will still
count it as a true value.