Added priorities for messages, pings at priority "high".
Messages can now (optionally) be sent at three different priority levels:
* low - messages are sent out only after all normal and high priority messages
	are sent.
* normal - messages are queued, throttled and sent as before.
* high - messages are sent immediately, without queueing or throttling.  The
	next "normal" or "low" priority message will be delayed further to
	to compensate.

Defaults to "normal", unless we are not registered yet in which case the default
  is "high" (so that registration is not throttled).

Channel joining and nickserv identify are now sent at "high" priority.

Pings (in response to the !ping command) are now sent at high priority so that
  the ping time is not affected by throttling.

Feed responses (in response to feed commands such as "!twitter") are now sent
at "low" priority so that the long response does not delay the sending of other
messages.

All other messages are now specifically sent at "normal" priority.