Added text-transform usertag plus a few filters.
Added text-transform usertag plus some filters used to convert and namecase
text directly in a db table.  This was done for a client so the filters and
tags are customized to what he wanted, but he also wanted me to share them with
the community in case they are helpful so here they are.