Added nick-based message queues.
The old message queue system had the vulnerability that one person could DOS the
bot output by loading up the message queue.  This has now been mitigated by the
addition of a new nick-based queue system.  There are now two message queues for
each nick and each channel so that messages are sent on a rotating basis to
nicks and channels instead of on a first come first served basis.  This means
that one person can slow down the overall message delivery but can't delay it
significantly as before.