Imported Oracle's centos2ol.sh script as ol2centos.sh
This is the script that Oracle likes to get people to use to switch over from
CentOS.  Since they released it under the Apache license 2.0 we can pretty much
use it how we want, so why not modify it to work in reverse?  We can use
Oracle's own script against them, so to speak.  Poetic justic if I ever saw it.